# openapi/v1.yaml
openapi: 3.0.3
info:
  title: Procezo API
  version: 1.4.0
paths:
...
  /api/v1/dossiers/:
    get:
      operationId: dossiers_list
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -created_at
          - -reference
          - created_at
          - reference
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: reference
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
      - in: query
        name: unit
        schema:
          type: string
          format: uuid
      tags:
      - dossiers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCaseList'
          description: ''
    post:
      operationId: dossiers_create
      tags:
      - dossiers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        reference:
          type: string
          readOnly: true
        unit:
          type: string
          format: uuid
          readOnly: true
        unit_code:
          type: string
          readOnly: true
        classification:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        assignee:
          type: integer
          readOnly: true
        assignee_username:
          type: string
          readOnly: true
        next_action:
          type: string
          readOnly: true
        version:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - assignee
      - assignee_username
      - classification
      - created_at
      - id
      - next_action
      - reference
      - status
      - unit
      - unit_code
      - updated_at
      - version
...
    CaseCreate:
      type: object
      properties:
        unit:
          type: string
          format: uuid
        classification:
          $ref: '#/components/schemas/ClassificationEnum'
        assignee:
          type: integer
          minimum: 1
        next_action:
          type: string
          maxLength: 240
        assignment_reason:
          type: string
          maxLength: 500
      required:
      - assignee
      - assignment_reason
      - classification
      - next_action
      - unit
...
    ClassificationEnum:
      enum:
      - 0
      - 1
      type: integer
      description: |-
        * `0` - 0
        * `1` - 1
...
    PaginatedCaseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Case'
...
    StatusEnum:
      enum:
      - open
      - in_progress
      type: string
      description: |-
        * `open` - Ouvert
        * `in_progress` - En cours
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid